requirements.txt
aiohttp==3.11.15
beautifulsoup4==4.13.3
Flask==3.1.0
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
matplotlib==3.10.1
numpy==2.0.2
pillow==11.1.0
scikit-learn==1.6.1
tensorflow==2.18.0
Werkzeug==3.1.3